requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements.in
#
attrs==19.3.0             # via automat, service-identity, twisted
automat==20.2.0           # via twisted
beautifulsoup4==4.9.1     # via -r requirements.in
blis==0.2.4               # via spacy, thinc
boto3==1.14.54            # via neuralcoref
botocore==1.17.54         # via boto3, s3transfer
cachetools==4.1.1         # via textacy
certifi==2020.6.20        # via requests
cffi==1.14.0              # via cryptography
chardet==3.0.4            # via requests
click==7.1.2              # via nltk
colorama==0.4.3           # via crayons
configparser==5.0.0       # via webdriver-manager
constantly==15.1.0        # via twisted
crayons==0.3.1            # via webdriver-manager
cryptography==2.9.2       # via pyopenssl, scrapy, service-identity
cssselect==1.1.0          # via parsel, scrapy
cymem==2.0.3              # via preshed, spacy, thinc
cytoolz==0.10.1           # via textacy
decorator==4.4.2          # via networkx
docutils==0.15.2          # via botocore
https://github.com/explosion/spacy-models/releases/download/en_core_web_lg-2.1.0/en_core_web_lg-2.1.0.tar.gz#egg=en_core_web_lg  # via -r requirements.in
hyperlink==19.0.0         # via twisted
idna==2.10                # via hyperlink, requests
incremental==17.5.0       # via twisted
itemadapter==0.1.0        # via scrapy
jellyfish==0.8.2          # via textacy
jmespath==0.10.0          # via boto3, botocore
joblib==0.16.0            # via nltk, scikit-learn, textacy
jsonschema==2.6.0         # via spacy
lxml==4.5.2               # via parsel, scrapy
murmurhash==1.0.2         # via spacy, thinc
networkx==2.5             # via textacy
neuralcoref==4.0          # via -r requirements.in
nltk==3.5                 # via -r requirements.in
nose==1.3.7               # via textsplit
numpy==1.19.0             # via blis, neuralcoref, pandas, pyemd, scikit-learn, scipy, spacy, textacy, textsplit, thinc
pandas==1.1.1             # via -r requirements.in
parsel==1.6.0             # via scrapy
plac==0.9.6               # via spacy, thinc
preshed==2.0.1            # via spacy, thinc
protego==0.1.16           # via scrapy
pyasn1-modules==0.2.8     # via service-identity
pyasn1==0.4.8             # via pyasn1-modules, service-identity
pycparser==2.20           # via cffi
pydispatcher==2.0.5       # via scrapy
pyemd==0.5.1              # via textacy
pyhamcrest==2.0.2         # via twisted
pyopenssl==19.1.0         # via scrapy
pyphen==0.9.5             # via textacy
python-aiml==0.9.3        # via -r requirements.in
python-dateutil==2.8.1    # via botocore, pandas
pytz==2020.1              # via pandas
queuelib==1.5.0           # via scrapy
regex==2020.6.8           # via nltk
requests==2.24.0          # via neuralcoref, spacy, textacy, webdriver-manager
s3transfer==0.3.3         # via boto3
scikit-learn==0.23.1      # via -r requirements.in, textacy
scipy==1.5.1              # via scikit-learn, textacy
scrapy==2.2.0             # via -r requirements.in
selenium==3.141.0         # via -r requirements.in
service-identity==18.1.0  # via scrapy
six==1.15.0               # via automat, cryptography, parsel, protego, pyopenssl, python-dateutil, w3lib
soupsieve==2.0.1          # via beautifulsoup4
spacy==2.1.0              # via -r requirements.in, en-core-web-lg, neuralcoref, textacy
srsly==1.0.2              # via spacy, textacy, thinc
textacy==0.9.1            # via -r requirements.in
textsplit==0.5            # via -r requirements.in
thinc==7.0.8              # via spacy
threadpoolctl==2.1.0      # via scikit-learn
toolz==0.10.0             # via cytoolz
tqdm==4.47.0              # via nltk, textacy, thinc
twisted==20.3.0           # via scrapy
urllib3==1.25.9           # via botocore, requests, selenium
w3lib==1.22.0             # via parsel, scrapy
wasabi==0.7.0             # via spacy, thinc
webdriver-manager==3.2.1  # via -r requirements.in
zope.interface==5.1.0     # via scrapy, twisted

# The following packages are considered to be unsafe in a requirements file:
# setuptools
